qcacmn: Compute freq for channels in other 6G power type
The API reg_compute_chan_to_freq() computes frequencies only for channels
present in master channel list. A 6G channel may not belong to the current
(current power type's) master channel but be present in different 6g power
type's master channel list.

So if a 6G channel is not found in the current (current power type's)
master channel list, then search for the channel in all the 6g power's
master channel list to compute the frequency.

Change-Id: I7ccdaeb1f9de8d9baabe59ad673d445ed36a6a2d