qcacmn: Fix flaw in cdp_update_mlo_ptnr_list() parameter
In mlo_ap_vdev_attach() we call cdp_update_mlo_ptnr_list(), and logic
exists which was supposed to test the return value for success.
However due to a misplaced parenthesis, the test for success is actually
applied to the last parameter, and that test result is sent as a
true/false value to cdp_update_mlo_ptnr_list().

To fix this issue relocate the parenthesis so that the last parameter
is correctly passed and the test for success is applied to the
return value.

Change-Id: I055150dae726a69f0e70f02f7fbe26550f48e558
CRs-Fixed: 3181220