qcacmn: Avoid timeout, in case active cmd fails to find valid req
If a new disconnect/connect req tries flush the pending
disconnect/connect request and while doing so, it
remove the CM pending request, but before it tried to
remove pending command from serialization, the command
becomes active in another thread.

This active command will try to fetch the CM request
which is removed and thus return without any handling.

This lead to active command timeout for the request.

To fix it remove the command from serialization active
queue, if disconnect/connect request was not found in
CM request list after the command become active.

Change-Id: I81d147060fb21d17e4aa82e99c073d1bd0c5db68
CRs-Fixed: 3212746