qcacmn: Return true from wlan_reg_is_afc_done if AFC/6GHz is not supported
When either of the macros CONFIG_AFC_SUPPORT or CONFIG_BAND_6GHZ are not
defined, the function wlan_reg_is_afc_done returned false by default.
As a result, the upper layer set all the 6GHz channels to NO_IR and the
vap was unable to come up.

Fix this by returning true from wlan_reg_is_afc_done when either of those
macros are not defined.

Change-Id: I4acb564023d2a2b318ba1497fd858a02987c117a
CRs-Fixed: 3214096