qcacmn: Increase the max rx descriptors to 16K
Currently the max number of rx descriptors, via INI,
is 4096, which is not sufficient if the number of
entries in rx buffer refill ring is increased.

Hence the set the max allowed number of rx descriptors
to 16K, which is also the max allowed entries in the
rx buffers refill ring.

Change-Id: I287a3f47525179ebf8de65e2f8310a961881916b
CRs-Fixed: 3224590