qcacmn: Add return value for qdf trace rate limited func
Add return value for qdf trace rate limited function to give
the caller a chance to know whether the current logging message
is bypassed or not. This allows the caller to skip additional
logging when logs are being rate limited.

Here is the value for this return:
false: log is not limited, output is normal.
true:  log is limted, will not be output in the log.

Change-Id: Ia5ae0e5676072142c3daf9b64a9bcf609692abf2
CRs-Fixed: 3284268